requirements.txt
kivy==2.3.0
buildozer~=1.5.0
cython~=3.0.10
opencv-python~=4.10.0.82
python-for-android~=2024.1.21
camera4kivy~=0.3.3
